Remove the spatial-navigation property
Whether spatnav should get invoked from arrow keys or something else
is for the UA to decide, and we don't need (at least in this level)
to provide control over it.

This also lets us get rid of the monkey-patch on UI-Events.

Closes #35

Also closes #22, #24, and #26 (as invalid / out of scope) by
deleting the section these issues applied to.