Change finding candidates into finding focusable areas
By moving the stepts that take the direction and starting point into
account from finding candidates to selecting the best one,
the "finding candidates" method can be made more generic.

This patches does that, and renames the relevant algorithm and API
accordingly.

This includes a convenience method to grab the nearest spatnav container
ancestor.

Closes #31