Improve the spatNavSearch API more precisely
There was unprecise behavior when the focus moves from focusable spatnav container to other focusables.

- Reason:
When using the spatNavSearch API, even if the candidates aren't included in the container, the best candidate was selected via selectBestCandidateFromEdge function.
- Fix:
When the focus moves from the focusable container to other focusables, there are two rules:
  - If the candidates are included in the container, the best candidate is selected via selectBestCandidateFromEdge  function
  - Else if all the candidates aren't included in the container, the best candidate is selected via selectBestCandidate function